Use size_t for more hash implementations
These all use hash functions that can return a 64bit value on
64bit platforms, so let's leverage that.

There are some places where we do not do this though, e.g. those
types that use the uint index as hash, such as for IndexedString
and those types that use IndexedString internally.

+++ kdevplatform/language/duchain/declarationid.h
@@ -111,7 +111,7 @@ public:
      *          depending on whether the id is direct or indirect,
      *          and thus you cannot compare hashes for declaration equality (use operator==() instead)
      */
-    uint hash() const
+    size_t hash() const
     {
         if (m_isDirect)
             return KDevHash() << m_directData.hash() << m_specialization.index();
